refactor: SelectBox 로직 분리
